requirements.txt
dj-database-url==2.2.0
Django==5.1
django-extensions==3.2.3
djangorestframework==3.15.2
django-seed==0.3.1
gunicorn==23.0.0
ipython==8.26.0
openpyxl==3.1.5
pandas==2.2.2
python-decouple==3.8
pytest==8.3.2
pytest-django==4.8.0
reportlab==4.2.2